requirements.txt
streamlit==1.24.0
python-pptx==0.6.21
pathlib==1.0.1
